Fix ignored file

+++ build.sh
@@ -4,20 +4,30 @@ FOLDERS="array bayes  genome  intruder  kmeans  labyrinth  ssca2  vacation  yada
 
 rm lib/*.o || true
 
+name=$1
+alias=$2
+
 cd ssync;
 make clean;
-make LOCK_VERSION="-DUSE_$2";
+make LOCK_VERSION="-DUSE_$alias";
+rc=$?
+if [[ $rc != 0 ]] ; then
+     echo ""
+     echo "=================================== ERROR BUILDING $name $alias ==================================="
+     echo ""
+     exit 1
+fi
 cd ..;
 
 for F in $FOLDERS
 do
     cd $F
     rm *.o || true
-    make -f Makefile.$1 LOCK_VERSION="-DUSE_$2"
+    make -f Makefile.$name LOCK_VERSION="-DUSE_$alias"
     rc=$?
     if [[ $rc != 0 ]] ; then
 	echo ""
-        echo "=================================== ERROR BUILDING $F $1 $2 ===================================="
+        echo "=================================== ERROR BUILDING $F $name $alias ===================================="
 	echo ""
         exit 1
     fi
